Add validate logic to block forced access to pages

+++ app.py
@@ -55,9 +55,12 @@ def get_event():
     Sort in ascending date order
     Call delete expired dates function """
 
-    events = list(mongo.db.events.find().sort("date", 1))
-    delete_expired_events()
-    return render_template("supper-club.html", events=events)
+    if session:
+        events = list(mongo.db.events.find().sort("date", 1))
+        delete_expired_events()
+        return render_template("supper-club.html", events=events)
+    else:
+        return redirect(url_for("home"))
 
 
 # Events search functionality
@@ -152,16 +155,19 @@ def profile(username):
     Call delete expired events function
     Return profile page with events, username and image """
 
-    username = mongo.db.users.find_one(
-        {"username": session["user"]})["username"]
-    user_profile = mongo.db.users.find_one({"username": session["user"]})
-    events = list(mongo.db.events.find({"created_by": session["user"]}))
+    if session:
+        username = mongo.db.users.find_one(
+            {"username": session["user"]})["username"]
+        user_profile = mongo.db.users.find_one({"username": session["user"]})
+        events = list(mongo.db.events.find({"created_by": session["user"]}))
 
-    if session["user"]:
-        delete_expired_events()
-        return render_template("profile.html",
-            username=username, events=events,
-            user_profile=user_profile)
+        if session["user"]:
+            delete_expired_events()
+            return render_template("profile.html",
+                username=username, events=events,
+                user_profile=user_profile)
+    else:
+        return redirect(url_for("home"))
 
     return redirect(url_for("sign_in"))
 
@@ -175,29 +181,32 @@ def edit_profile(user_id):
     Update the users current info in DB with the new info
     Else block the action and display a message to the user """
 
-    session_user = mongo.db.users.find_one(
-        {"username": session["user"]})["username"]
-    users = list(mongo.db.users.find({"_id": ObjectId(user_id)}))
+    if session:
+        session_user = mongo.db.users.find_one(
+            {"username": session["user"]})["username"]
+        users = list(mongo.db.users.find({"_id": ObjectId(user_id)}))
 
-    for user in users:
-        username = user['username']
+        for user in users:
+            username = user['username']
 
-        if session_user == username:
-            if request.method == "POST":
-                update_profile = {
-                    "username": request.form.get("username").lower(),
-                    "password": generate_password_hash(
-                        request.form.get("password")),
-                    "profile_pic": request.form.get("profile_pic")
-                }
-                session["user"] = request.form.get("username").lower()
-                mongo.db.users.update(
-                    {"_id": ObjectId(user_id)}, update_profile)
-                flash("Profile Updated")
+            if session_user == username:
+                if request.method == "POST":
+                    update_profile = {
+                        "username": request.form.get("username").lower(),
+                        "password": generate_password_hash(
+                            request.form.get("password")),
+                        "profile_pic": request.form.get("profile_pic")
+                    }
+                    session["user"] = request.form.get("username").lower()
+                    mongo.db.users.update(
+                        {"_id": ObjectId(user_id)}, update_profile)
+                    flash("Profile Updated")
+                    return redirect(url_for("profile", username=session['user']))
+            else:
+                flash("You Do Not Have Permission To Perform This Action")
                 return redirect(url_for("profile", username=session['user']))
-        else:
-            flash("You Do Not Have Permission To Perform This Action")
-            return redirect(url_for("profile", username=session['user']))
+    else:
+        return redirect(url_for("home"))
 
     return render_template("edit-profile.html", user=user)
 
@@ -211,21 +220,24 @@ def delete_profile(user_id):
     Else block action and display a message to the user
     Remove the deleted user from the session """
 
-    session_user = mongo.db.users.find_one(
-        {"username": session["user"]})["username"]
-    users = list(mongo.db.users.find({"_id": ObjectId(user_id)}))
+    if session:
+        session_user = mongo.db.users.find_one(
+            {"username": session["user"]})["username"]
+        users = list(mongo.db.users.find({"_id": ObjectId(user_id)}))
 
-    for user in users:
-        username = user['username']
+        for user in users:
+            username = user['username']
 
-        if session_user == username:
-            mongo.db.events.delete_many({"created_by": session["user"]})
-            mongo.db.users.delete_one({"_id": ObjectId(user_id)})
-            session.pop("user")
-            return redirect(url_for("home"))
-        else:
-            flash("You Do Not Have Permission To Perform This Action")
-            return redirect(url_for("profile", username=session['user']))
+            if session_user == username:
+                mongo.db.events.delete_many({"created_by": session["user"]})
+                mongo.db.users.delete_one({"_id": ObjectId(user_id)})
+                session.pop("user")
+                return redirect(url_for("home"))
+            else:
+                flash("You Do Not Have Permission To Perform This Action")
+                return redirect(url_for("profile", username=session['user']))
+    else:
+        return redirect(url_for("home"))
 
 
 # Sign user out
@@ -234,9 +246,12 @@ def sign_out():
     """ Remove session user
     Return sign in page """
 
-    flash("You have been logged out")
-    session.pop("user")
-    return redirect(url_for("sign_in"))
+    if session:
+        flash("You have been logged out")
+        session.pop("user")
+        return redirect(url_for("sign_in"))
+    else:
+        return redirect(url_for("home"))
 
 
 # Contact page
@@ -257,26 +272,29 @@ def create_event():
     Find and sort category name order
     Redirect to get event page """
 
-    if request.method == "POST":
-        event = {
-            "event_name": request.form.get("event_name"),
-            "location": request.form.get("location"),
-            "category_name": request.form.get("category_name"),
-            "date": request.form.get("date"),
-            "starter": request.form.get("starter"),
-            "main": request.form.get("main"),
-            "dessert": request.form.get("dessert"),
-            "extras": request.form.get("extras"),
-            "counter": int(request.form.get("counter")),
-            "created_by": session["user"]
-        }
-        mongo.db.events.insert_one(event)
-        flash("Supper Club Successfully Added")
-        delete_expired_events()
-        return redirect(url_for("get_event"))
+    if session:
+        if request.method == "POST":
+            event = {
+                "event_name": request.form.get("event_name"),
+                "location": request.form.get("location"),
+                "category_name": request.form.get("category_name"),
+                "date": request.form.get("date"),
+                "starter": request.form.get("starter"),
+                "main": request.form.get("main"),
+                "dessert": request.form.get("dessert"),
+                "extras": request.form.get("extras"),
+                "counter": int(request.form.get("counter")),
+                "created_by": session["user"]
+            }
+            mongo.db.events.insert_one(event)
+            flash("Supper Club Successfully Added")
+            delete_expired_events()
+            return redirect(url_for("get_event"))
 
-    categories = mongo.db.categories.find().sort("category_name", 1)
-    return render_template("create-event.html", categories=categories)
+        categories = mongo.db.categories.find().sort("category_name", 1)
+        return render_template("create-event.html", categories=categories)
+    else:
+        return redirect(url_for("home"))
 
 
 # Update event guest counter and email addresses
@@ -313,37 +331,40 @@ def edit_event(event_id):
     Else block the action and display a message
     Redirect user to profile page """
 
-    username = mongo.db.users.find_one(
-        {"username": session["user"]})["username"]
-    events = list(mongo.db.events.find({"_id": ObjectId(event_id)}))
+    if session:
+        username = mongo.db.users.find_one(
+            {"username": session["user"]})["username"]
+        events = list(mongo.db.events.find({"_id": ObjectId(event_id)}))
 
-    for event in events:
-        author = event['created_by']
+        for event in events:
+            author = event['created_by']
 
-        if username == author:
-            if request.method == "POST":
-                submit = {
-                    "event_name": request.form.get("event_name"),
-                    "location": request.form.get("location"),
-                    "category_name": request.form.get("category_name"),
-                    "date": request.form.get("date"),
-                    "starter": request.form.get("starter"),
-                    "main": request.form.get("main"),
-                    "dessert": request.form.get("dessert"),
-                    "extras": request.form.get("extras"),
-                    "counter": int(request.form.get("counter")),
-                    "created_by": session["user"]
-                }
-                mongo.db.events.update({"_id": ObjectId(event_id)}, submit)
-                flash("Supper Club Successfully Updated")
+            if username == author:
+                if request.method == "POST":
+                    submit = {
+                        "event_name": request.form.get("event_name"),
+                        "location": request.form.get("location"),
+                        "category_name": request.form.get("category_name"),
+                        "date": request.form.get("date"),
+                        "starter": request.form.get("starter"),
+                        "main": request.form.get("main"),
+                        "dessert": request.form.get("dessert"),
+                        "extras": request.form.get("extras"),
+                        "counter": int(request.form.get("counter")),
+                        "created_by": session["user"]
+                    }
+                    mongo.db.events.update({"_id": ObjectId(event_id)}, submit)
+                    flash("Supper Club Successfully Updated")
+                    return redirect(url_for("profile", username=session['user']))
+            else:
+                flash("You Do Not Have Permission To Perform This Action")
                 return redirect(url_for("profile", username=session['user']))
-        else:
-            flash("You Do Not Have Permission To Perform This Action")
-            return redirect(url_for("profile", username=session['user']))
 
-    event = mongo.db.events.find_one({"_id": ObjectId(event_id)})
-    categories = mongo.db.categories.find().sort("category_name", 1)
-    return render_template("edit-event.html", event=event, categories=categories)
+        event = mongo.db.events.find_one({"_id": ObjectId(event_id)})
+        categories = mongo.db.categories.find().sort("category_name", 1)
+        return render_template("edit-event.html", event=event, categories=categories)
+    else:
+        return redirect(url_for("home"))
 
 
 # Delete event
@@ -354,28 +375,23 @@ def delete_event(event_id):
     Else block the action and display a message to the user
     Return user to profile """
 
-    username = mongo.db.users.find_one(
-        {"username": session["user"]})["username"]
-    events = list(mongo.db.events.find({"_id": ObjectId(event_id)}))
+    if session:
+        username = mongo.db.users.find_one(
+            {"username": session["user"]})["username"]
+        events = list(mongo.db.events.find({"_id": ObjectId(event_id)}))
 
-    for event in events:
-        author = event['created_by']
+        for event in events:
+            author = event['created_by']
 
-        if username == author:
-            mongo.db.events.delete_one({"_id": ObjectId(event_id)})
-            flash("Supper Club Successfully Deleted")
-        else:
-            flash("You Do Not Have Permission To Perform This Action")
+            if username == author:
+                mongo.db.events.delete_one({"_id": ObjectId(event_id)})
+                flash("Supper Club Successfully Deleted")
+            else:
+                flash("You Do Not Have Permission To Perform This Action")
 
-    return redirect(url_for("profile", username=session['user']))
-
-
-# Supper club
-@app.route("/supper-club")
-def supper_club():
-    """ Render supper club page """
-
-    return render_template("supper-club.html")
+        return redirect(url_for("profile", username=session['user']))
+    else:
+        return redirect(url_for("home"))
 
 
 # Custom error pages
@@ -425,4 +441,4 @@ def server_response_error(e):
 if __name__ == "__main__":
     app.run(host=os.environ.get("IP"),
             port=int(os.environ.get("PORT")),
-            debug=False)
+            debug=True)
